gcp_function/requirements.txt
functions-framework>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-vision>=3.0.0
vertexai>=1.0.0